free creativity through personality.lua
Letting the community play around with the AIC values was always a lot of fun and new tricks and giggles have been discovered that proved useful in certain scenarios. Even if nieche we shoudn't restrict the creativity of the community, letting the AI explore some 'cheaty' methods of gathering resources through market mechanics. Therefore this lifts the restrictions on maximum recources, tradeamounts and buildings by letting the modder bypass the negative values and exploring the full r ange of opportunities given with these ranges of integers.